Inline EU4 compression into wasm-eu4
The compression wasm bundle was initially separated from the main bundle
as it used brotli, whose heavy encoder significantly increased the
bundle size.

Now there seems very little reason to keep them separate.
