--- src/app/app/features/compress/index.tsx
@@ -1,1 +1,0 @@
-export * from "./compress-hooks";
