+++ src/wasm-compress/Cargo.toml
@@ -18,10 +18,5 @@ zstd_rust = ["pdx-zstd/zstd_rust"]
 zstd_c_fat_lto = ["pdx-zstd/zstd_c_fat_lto"]
 
 [dependencies]
-flate2 = { default-features = false, features = ["zlib-rs"] , workspace = true }
-js-sys = { workspace = true }
 pdx-zstd = { workspace = true }
-rawzip = { workspace = true }
-serde = { features = ["derive"] , workspace = true }
-tsify = { default-features = false, features = ["js"] , workspace = true }
 wasm-bindgen = { workspace = true }
