+++ src/wasm-compress/tests/test.rs
@@ -1,46 +1,12 @@
-use rawzip::CompressionMethod;
-use std::io::Read;
-use wasm_compress::download_transformation;
-
-fn compress(data: Vec<u8>) -> Vec<u8> {
-    let c = wasm_compress::init_compression(data);
-
-    c.compress_cb(None).unwrap()
-}
+use std::io::Write;
+use wasm_compress::decode_zstd;
 
 #[test]
-fn test_recompression_plaintext() {
+fn test_decode_zstd_roundtrip() {
     let data = b"hello world";
-    let compressed = compress(data.to_vec());
-    let actual = pdx_zstd::decode_all(&compressed).unwrap();
-    assert_eq!(actual.as_slice(), &data[..]);
-}
-
-#[test]
-fn test_recompression_zip() {
-    let data = include_bytes!("test.zip");
-    let compressed = compress(data.to_vec());
-    let archive = rawzip::ZipArchive::from_slice(compressed.as_slice()).unwrap();
-    let mut entries = archive.entries();
-    let entry = entries.next_entry().unwrap().unwrap();
-    assert_eq!(entry.file_path().as_ref(), b"test.txt");
-    assert_eq!(entry.compression_method(), CompressionMethod::Zstd);
-    let file = archive.get_entry(entry.wayfinder()).unwrap();
-    let actual = pdx_zstd::decode_all(file.data()).unwrap();
-    assert_eq!(actual.as_slice(), b"aaaaaaaaaa\n");
-    assert!(entries.next_entry().unwrap().is_none());
-
-    let original = download_transformation(compressed).unwrap();
-    let archive = rawzip::ZipArchive::from_slice(original.as_slice()).unwrap();
-    let mut entries = archive.entries();
-    let entry = entries.next_entry().unwrap().unwrap();
-    assert_eq!(entry.file_path().as_ref(), b"test.txt");
-    assert_eq!(entry.compression_method(), CompressionMethod::Deflate);
-    let file = archive.get_entry(entry.wayfinder()).unwrap();
-    let mut buf = Vec::new();
-    flate2::bufread::DeflateDecoder::new(file.data())
-        .read_to_end(&mut buf)
-        .unwrap();
-    assert_eq!(buf.as_slice(), b"aaaaaaaaaa\n");
-    assert!(entries.next_entry().unwrap().is_none());
+    let mut encoder = pdx_zstd::Encoder::new(Vec::new(), 7).unwrap();
+    encoder.write_all(data).unwrap();
+    let compressed = encoder.finish().unwrap();
+    let decoded = decode_zstd(&compressed).unwrap();
+    assert_eq!(decoded.as_slice(), &data[..]);
 }
